requirements.txt
Django==5.2.1
djangorestframework==3.15
psycopg2-binary
python-dotenv==1.1.0
djangorestframework-simplejwt
django-filter
drf-spectacular
pytest
pytest-django